// server/enums/inductionScheduleStatusValue.ts
enum InductionScheduleStatusValue {
  SCHEDULED = 'SCHEDULED',
  COMPLETE = 'COMPLETE',
  EXEMPT_PRISONER_DRUG_OR_ALCOHOL_DEPENDENCY = 'EXEMPT_PRISONER_DRUG_OR_ALCOHOL_DEPENDENCY',
  EXEMPT_PRISONER_OTHER_HEALTH_ISSUES = 'EXEMPT_PRISONER_OTHER_HEALTH_ISSUES',
  EXEMPT_PRISONER_FAILED_TO_ENGAGE = 'EXEMPT_PRISONER_FAILED_TO_ENGAGE',
  EXEMPT_PRISONER_ESCAPED_OR_ABSCONDED = 'EXEMPT_PRISONER_ESCAPED_OR_ABSCONDED',
  EXEMPT_PRISONER_SAFETY_ISSUES = 'EXEMPT_PRISONER_SAFETY_ISSUES',
  EXEMPT_PRISON_REGIME_CIRCUMSTANCES = 'EXEMPT_PRISON_REGIME_CIRCUMSTANCES',
  EXEMPT_PRISON_STAFF_REDEPLOYMENT = 'EXEMPT_PRISON_STAFF_REDEPLOYMENT',
  EXEMPT_PRISON_OPERATION_OR_SECURITY_ISSUE = 'EXEMPT_PRISON_OPERATION_OR_SECURITY_ISSUE',
  EXEMPT_SECURITY_ISSUE_RISK_TO_STAFF = 'EXEMPT_SECURITY_ISSUE_RISK_TO_STAFF',
  EXEMPT_SYSTEM_TECHNICAL_ISSUE = 'EXEMPT_SYSTEM_TECHNICAL_ISSUE',
  EXEMPT_PRISONER_TRANSFER = 'EXEMPT_PRISONER_TRANSFER',
  EXEMPT_PRISONER_RELEASE = 'EXEMPT_PRISONER_RELEASE',
  EXEMPT_PRISONER_DEATH = 'EXEMPT_PRISONER_DEATH',
}

export default InductionScheduleStatusValue